+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Tested on OS build 22000.376.
   * Fixed a bug that displayed wrong window previews when the combine taskbar buttons option was set to never combine (#564) (.17)
   * Possibility to set position on screen (top/bottom) from the Properties UI
 * Restoring default settings only asks for elevation if required (for the moment, only if you have registered the application as shell extension) (.18)
+* Fixed the context menu not working (and a potential associated crash) of the new Microsoft IME (#598, #588) (.19) (huge thanks to @Simplestas)
 
 
 #### Simple Window Switcher